23.16

+++ database/migrations/2022_11_19_125134_add_role_id_to_users.php
@@ -14,10 +14,10 @@ class AddRoleIdToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            
+            if (!Schema::hasColumn('users', 'role_id')) {
                 $table->integer('role_id')->unsigned()->default(2);
                 $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            
+            }
         });
     }
